Changing to ubuntu base image
So that I can install go 1.11beta3. We need go 1.11beta3 so that the proxy can use it to fill its caches on the backend. It's technically possible to use alpine to do all that stuff but it's harder b/c you need to configure Go to build with musl