[refs #730] Include dev meeting in docs
The Athens development meeting happens weekly but nobody knows where to
find it.. I felt that the current organisation of these pages wasn't
easily readable so have sorted the community style pages separate from
those on say using Git or GitHub. I took a lot of the content from the
document so I altered it slightly so it fits with the voice of the
current document.

It was suggested that we move to leaving out the timezone as we'd have
to alter with DST twice a year.

Also moved Maintainers > maintainers as suggested.

Reading back 'to watch back anytime' does sound a bit strange. Simply
removed.

Signed-off-by: Chris M <millscj01@gmail.com>