Add to docker-compose and standardise the file
Whilst we wait for a ticket to be completed we need to star the initial
part of this work. I've updated the compose file to include azurite and
also sorted that file so it's standard in some way, we used various ways
to define environment variables/commands and things were a bit all over
the place.

Signed-off-by: Chris M <millscj01@gmail.com>