Add Dockerfile for docs
There's no official image for hugo, so lets make our own and have the
Makefile build the Dockerfile

Add some basic documentation

I've added some simple documentation and tidied up the Dockerfile. I
move the file to use best practice and reduced the size slightly.
Personally I don't think we'll need to alter the address and not watch
files as it's only going to be used for local development atm. It might
be useful in future to move to being able to alter these if we're using
this in production

Don't use docs use hugo

It would be confusing for people to docker pull the docs but not have
any docs included in them. hugo as the name would make way more sense.

Don't lock down to digest

We should trust official libs, especially the alpine build.

Signed-off-by: Chris M <millscj01@gmail.com>