pkg/module: include SSH_AUTH_SOCK in go get env (#1184)
* pkg/module: include SSH_AUTH_SOCK in go get env

When running locally, it's convenient to use a user's ssh-agent. For
that to work, we need to propagate SSH_AUTH_SOCK.

* pkg/module: Verify existence for SSH_AUTH_SOCK

SSH_AUTH_SOCK is expected to indicate a unix socket on the local
filesystem. If it isn't a unix socket (after symlink resolution) we can
ignore it. (although ssh will ignore it as well)