Add ability to annotate the proxy Service via Helm
It is often use to add custom annotations to a Service deployed
through Helm, so this PR adds the ability for a user to specify
annotations in their values file. For such use case we have seen
is when working in conjunction with external-dns, we want to assign
DNS properties of the Service, but currently cannot because this
Helm chart does not support it.