Install git-lfs on proxy container
For repositories that store some files in git-lfs, the checksum will
differ for clients with and without git-lfs. Clients without git-lfs
will use the git-lfs pointer [0] for calculation, clients with will use
the real files.
This adds git-lfs to the container to match expected behaviour and
resolve checksum mismatches in these cases.

[0] https://github.com/git-lfs/git-lfs/blob/master/docs/spec.md