Adding a better error message for download mode
When it's missing, we write a short message saying so, and when it's invalid, we do approximately the same thing. Athens has a lot of configuration options, so it's naturally easy to get them wrong or forget something. Nicer error messages should help.

If this code is acceptable, I'd like to apply it to a few other config values (in a follow-up PR) that are easy to get wrong.

ref https://github.com/gomods/athens/issues/1491, since this checks existence and validates values for download mode, and also reports errors in a more human-friendly way

cc/ @bluekeyes