Improving the project status section of our README
Athens is not really "beta" anymore because teams are using it and we are already taking backward compatibility into consideration actively.

I believe that the version number is a different but related discussion that we should have.

I've also updated the contributing documentation and removed the section about dev meetings