Fix download mode "async_redirect" not caching modules
Given the following setup:

* ATHENS_SINGLE_FLIGHT_TYPE="redis"
* ATHENS_STORAGE_TYPE="s3"

We observed that any mod using "async_redirect" was never cached in s3.

The issue was that the `go stasher.Stash` call passed in the context
created from the HTTP handler, where the context is canceled once the
request has been served to the client.

The consequence of the context being canceled after redirect (async_redirect)
or error (async) was that the s3 client would see the context as canceled
and just return an error immediately, preventing the caching flow to
complete.

It's worth noting that the following setup do not have the problem:

* ATHENS_SINGLE_FLIGHT_TYPE="memory"
* ATHENS_STORAGE_TYPE="s3"

Because the memory (default) in-flight type do not call "checker.Exists"
which have the offending code.

The proposed change will retain tracing behaviour correctly, and any
`WithValue()` calls set before or after

Signed-off-by: Christian Winther <jippignu@gmail.com>