Handle SIGTERM on unix-like OS
When running Athens on a unix-like OS (linux, darwin, etc), the TERM
signal is not being handled to perform a graceful shutdown, only INT.
This is especially problematic on Kubernetes, where the default signal
sent to pods is TERM.

This commit adds a new package to get the appropriate signals based on
the operating system, using go:build tags. For 'unix', it includes
os.Interrupt AND syscall.SIGTERM, but only os.Interrupt for others,
which may not support that signal.